+++ src/resources/manifest.php
@@ -2,5 +2,5 @@
 
 return [
     'name'    => 'Vanilo Order Module',
-    'version' => '2.0.0'
+    'version' => '2.1.0'
 ];